src/requirements.txt
pandas
sqlalchemy
discord.py
requests
tomli
psycopg2-binary